docker: Only delete asc if exists

+++ docker/ci-build.sh
@@ -88,7 +88,9 @@ get_source() {
 
 	origtargz --clean
 	# Delete asc files after they have been verified
-	rm "../*.asc"
+	if [ -f "../*.asc" ]; then
+		rm "../*.asc"
+	fi
 	origtargz --unpack=yes
 }
